Initial port commit

+++ src/entities/enemy_arrays.h
@@ -1,20 +1,20 @@
 #include "../images/enemy_sprites.h"
 
-const int enemy_array[] = {
-//  max_hp,     hp,         atk_low,   atk_high,    exp,        (int) sprite
-	20,         20,        	5,         10,          20,         (int) ghost,
-	30,        	30,         6,         10,          22,         (int) fireKid,
-	40,         40,       	7,         11,          24,         (int) frog,
-    45,         45,         8,         11,          30,         (int) doomSlime,
-    55,         55,         9,         12,          35,         (int) floatingEye,
-    65,         65,         5,         16,          40,         (int) mrEye,
-	75,        	75,         10,        19,          50,         (int) fetus,
-    95,         95,         11,        20,          60,         (int) spooks,
-    105,        105,        13,        22,          75,         (int) skeleton,
-    120,        120,        15,        25,          85,         (int) flower_man,
-    130,        130,        16,        26,          95,         (int) alien,
-    150,        150,        17,        30,          100,        (int) alienKing,
-    160,        160,        20,        33,          110,        (int) evil_thing,
-    180,        180,        22,        35,          120,        (int) oni,
-    200,        200,       	24,        35,          140,        (int) knight
+const long enemy_array[] = {
+//  max_hp,     hp,         atk_low,   atk_high,    exp,        (long) sprite
+	20,         20,        	5,         10,          20,         (long) ghost,
+	30,        	30,         6,         10,          22,         (long) fireKid,
+	40,         40,       	7,         11,          24,         (long) frog,
+    45,         45,         8,         11,          30,         (long) doomSlime,
+    55,         55,         9,         12,          35,         (long) floatingEye,
+    65,         65,         5,         16,          40,         (long) mrEye,
+	75,        	75,         10,        19,          50,         (long) fetus,
+    95,         95,         11,        20,          60,         (long) spooks,
+    105,        105,        13,        22,          75,         (long) skeleton,
+    120,        120,        15,        25,          85,         (long) flower_man,
+    130,        130,        16,        26,          95,         (long) alien,
+    150,        150,        17,        30,          100,        (long) alienKing,
+    160,        160,        20,        33,          110,        (long) evil_thing,
+    180,        180,        22,        35,          120,        (long) oni,
+    200,        200,       	24,        35,          140,        (long) knight
 };
